Refactor assistant orchestration and split UI
Refactor the assistant runtime to break up box3_assistant.c into assistant boot, session, watchdog, and presence modules, move timer/weather/Hue command execution behind feature handlers, split ui_status.c by display/render/font responsibilities, and add host tests for the new registry and transport helpers.

Closes #8
