+++ main/CMakeLists.txt
@@ -1,27 +1,39 @@
 idf_component_register(
     SRCS
+        "assistant/boot.c"
+        "assistant/command_registry.c"
+        "assistant/presence.c"
+        "assistant/session.c"
+        "assistant/watchdog.c"
         "commands/assistant_command_dispatch.c"
         "commands/assistant_command_text.c"
         "board/board_audio.c"
         "board/ui_status.c"
+        "board/ui_status_display.c"
+        "board/ui_status_font.c"
+        "board/ui_status_render.c"
         "assistant_diagnostics.c"
         "assistant_state.c"
         "box3_assistant.c"
         "hue/hue_command_map.c"
+        "hue/hue_command_handler.c"
         "hue/hue_command_runtime.c"
         "hue/hue_client.c"
         "hue/hue_discovery_response.c"
         "hue/hue_group_store.c"
+        "net/http_trace.c"
         "net/line_socket.c"
         "stt/local_stt_client.c"
         "stt/local_stt_protocol.c"
         "system/time_support.c"
         "system/wifi_support.c"
+        "timer/timer_command_handler.c"
         "timer/timer_parse.c"
         "timer/timer_runtime.c"
         "tts/local_tts_client.c"
         "tts/tts_player.c"
         "weather/weather_client.c"
+        "weather/weather_command_handler.c"
         "weather/weather_format.c"
         "weather/weather_open_meteo_provider.c"
     INCLUDE_DIRS "."
