+++ tests/esp_err.h
@@ -1,0 +1,7 @@
+#pragma once
+
+typedef int esp_err_t;
+
+#define ESP_OK                0
+#define ESP_ERR_INVALID_ARG   0x102
+#define ESP_ERR_INVALID_STATE 0x103
