+++ tests/test_main.c
@@ -6,12 +6,18 @@ extern const test_case_t g_assistant_state_tests[];
 extern const int g_assistant_state_test_count;
 extern const test_case_t g_command_dispatch_tests[];
 extern const int g_command_dispatch_test_count;
+extern const test_case_t g_command_registry_tests[];
+extern const int g_command_registry_test_count;
 extern const test_case_t g_command_text_tests[];
 extern const int g_command_text_test_count;
 extern const test_case_t g_hue_discovery_response_tests[];
 extern const int g_hue_discovery_response_test_count;
+extern const test_case_t g_http_trace_tests[];
+extern const int g_http_trace_test_count;
 extern const test_case_t g_local_stt_protocol_tests[];
 extern const int g_local_stt_protocol_test_count;
+extern const test_case_t g_request_cancel_tests[];
+extern const int g_request_cancel_test_count;
 extern const test_case_t g_weather_format_tests[];
 extern const int g_weather_format_test_count;
 extern const test_case_t g_timer_parse_tests[];
@@ -41,9 +47,12 @@ static void run_test_cases(const test_case_t *tests, int count) {
 int main(void) {
     run_test_cases(g_assistant_state_tests, g_assistant_state_test_count);
     run_test_cases(g_command_dispatch_tests, g_command_dispatch_test_count);
+    run_test_cases(g_command_registry_tests, g_command_registry_test_count);
     run_test_cases(g_command_text_tests, g_command_text_test_count);
     run_test_cases(g_hue_discovery_response_tests, g_hue_discovery_response_test_count);
+    run_test_cases(g_http_trace_tests, g_http_trace_test_count);
     run_test_cases(g_local_stt_protocol_tests, g_local_stt_protocol_test_count);
+    run_test_cases(g_request_cancel_tests, g_request_cancel_test_count);
     run_test_cases(g_timer_parse_tests, g_timer_parse_test_count);
     run_test_cases(g_timer_runtime_tests, g_timer_runtime_test_count);
     run_test_cases(g_weather_format_tests, g_weather_format_test_count);
