+++ tests/weather/weather_command_handler.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "assistant/command_handler.h"
+
+const assistant_command_handler_t *weather_command_handler_get(void);
